+++ src/Enforce/runnables/rules/content/equals/index.js
@@ -1,0 +1,9 @@
+// @flow
+
+function equals(value: AnyValue, arg1: AnyValue): boolean {
+    return value === arg1;
+}
+
+equals.negativeForm = 'notEquals';
+
+export default equals;